LayoutTests/fast/writing-mode/vertical-inline-block-hittest.html

<!DOCTYPE html>
<html>
<body style="margin: 0px">
<div id="target" style="-webkit-writing-mode: vertical-rl; border: solid 1px black; font-family: Ahem">
<br />
<span style="display: inline-block"><a>b</a></span>
</div>
<p id="description"></p>
<div id="console"></div>
<script src='../js/resources/js-test-pre.js'></script>
<script>
function runTests() {
    description("Test that we can click 2nd line inside inline-block in vertical writing mode.");
    elementInsideInlineBlockInSecondLine = document.querySelector("a");
    shouldBe("document.elementFromPoint(5, 5)", "elementInsideInlineBlockInSecondLine");
    document.getElementById("target").style.display = "none";
}
runTests();
</script>
<script src='../js/resources/js-test-post.js'></script>
</body>
</html>
