safari-600.7-branch/LayoutTests/fast/writing-mode/vertical-inline-block-hittest-expected.txt

Test that we can click 2nd line inside inline-block in vertical writing mode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.elementFromPoint(5, 5) is elementInsideInlineBlockInSecondLine
PASS successfullyParsed is true

TEST COMPLETE
